implement antonia's timer
