--- demo-code/recipe-info-rough/recipe-page.css
@@ -1,297 +1,0 @@
-/* stylelint-disable selector-no-vendor-prefix */
-/* stylelint-disable selector-class-pattern */
-/* stylelint-disable selector-id-pattern */
-@import url("https://fonts.googleapis.com/css2?family=Lato&display=swap");
-
-html,
-body {
-  font-family: Lato, sans-serif;
-  margin: 0;
-  padding: 0;
-}
-
-/* Recipe info page CSS */
-
-.recipe-info-page {
-  background-color: white;
-  box-shadow: 0 0 10px rgb(0 0 0 / 15%);
-  margin: 5rem;
-  margin-bottom: 2.5rem;
-  margin-left: auto;
-  margin-right: auto;
-  max-width: 44rem;
-  padding: 15px;
-  transition: all 0.2s ease;
-  width: 80%;
-}
-
-/* generic button CSS style */
-
-.info-page-btn {
-  border: 1px solid rgb(0 0 0);
-  color: rgb(0 0 0);
-  cursor: pointer;
-  display: inline-block;
-  font-size: 14px;
-  padding: 5px 10px;
-  text-align: center;
-  text-decoration: none;
-}
-
-.back-btn {
-  background-color: #f1f7f7;
-}
-
-.org-btn {
-  background-color: #fff0d4;
-  width: auto;
-}
-
-.unit-btn {
-  background-color: #f1f7f7;
-  display: block;
-  font-size: large;
-  margin: auto;
-  width: 100%;
-}
-
-.measurement-btn {
-  background-color: #f1f7f7;
-  font-size: large;
-  margin-left: auto;
-  width: 50%;
-}
-
-/* div wrappers */
-
-.title-edit-wrapper {
-  align-items: center;
-  display: flex;
-  height: 2rem;
-  justify-content: space-between;
-  margin-top: 1rem;
-}
-
-.rating-favorite-wrapper {
-  align-items: center;
-  display: flex;
-  height: 2rem;
-  justify-content: space-between;
-  margin-bottom: 1rem;
-  margin-top: 0.5rem;
-}
-
-.steps-wrapper {
-  display: grid;
-  height: auto;
-  margin-bottom: 0.5rem;
-  margin-top: 1.25rem;
-  outline: 1px solid rgb(0 0 0);
-  overflow: scroll;
-  padding: 1rem;
-}
-
-.videos-wrapper {
-  display: grid;
-  margin-bottom: 0.5rem;
-  margin-top: 1.25rem;
-  outline: 1px solid rgb(0 0 0);
-  padding: 1rem;
-}
-
-.nutrition-wrapper {
-  display: grid;
-  margin-bottom: 0.5rem;
-  margin-top: 1.25rem;
-  outline: 1px solid rgb(0 0 0);
-  padding: 1rem;
-}
-
-.recipe-description-wrapper {
-  margin-bottom: 0.5rem;
-  margin-top: 0.5rem;
-  outline: 1px solid rgb(0 0 0);
-  padding: 2rem;
-}
-
-/* Ratings */
-
-.rating {
-  align-items: center;
-  display: inline-flex;
-}
-
-.rating-stars {
-  align-items: center;
-  display: flex;
-  height: 1rem;
-}
-
-.divider {
-  align-items: center;
-  background: rgb(0 0 0);
-  display: inline-block;
-  height: 1.5rem;
-  margin-left: 1rem;
-  margin-right: 1rem;
-  width: 1px;
-}
-
-/* sub-header CSS style */
-
-.recipe-info-body {
-  display: flex;
-  margin-top: 0.75rem;
-  outline: 1px solid rgb(0 0 0);
-  padding: 0.5rem;
-}
-
-.info-body-section {
-  display: grid;
-  margin-bottom: 0.5rem;
-  margin-top: 0.5rem;
-  outline: 0 solid rgb(0 0 0);
-  padding: 0.5rem;
-  width: 100%;
-}
-
-.ingredients-wrapper {
-  border-right: 1px solid rgb(0 0 0);
-  height: 26pc;
-  overflow: scroll;
-}
-
-.quick-facts-wrapper {
-  border-bottom: 1px solid rgb(0 0 0);
-  height: 13pc;
-}
-
-.measurements-wrapper {
-  height: 8pc;
-  padding-top: 1rem;
-}
-
-.serving-adjust {
-  align-items: center;
-  display: flex;
-  margin: auto;
-  width: 100%;
-}
-
-/* body CSS style */
-
-/* image CSS style */
-
-.recipe-img {
-  height: auto;
-  margin: 0;
-  max-height: 27rem;
-  object-fit: cover;
-  outline: 1px solid rgb(0 0 0);
-  width: 100%;
-}
-
-/* text CSS style */
-
-.info-title {
-  font-size: 24px;
-}
-
-.info-subtitle {
-  font-size: 20px;
-  margin: 0;
-  padding: 0;
-  text-decoration: underline;
-  text-decoration-color: gold;
-  text-decoration-thickness: 3px;
-  text-underline-position: under;
-}
-
-.description {
-  font-size: 16px;
-  margin: 0;
-}
-
-.serving-size {
-  margin: 1rem 2rem;
-}
-
-.subtitle.measure-title {
-  margin-bottom: 1rem;
-}
-
-/* list CSS style */
-
-.info-list {
-  padding: 0 1rem;
-}
-
-.rating-box {
-  align-items: center;
-  display: flex;
-  padding-left: 1rem;
-}
-
-.rating-stats {
-  align-items: center;
-  color: rgb(121 121 121);
-  display: flex;
-  font-size: 14px;
-  margin: 0;
-  padding: 0;
-}
-
-.info-facts,
-.info-ingredient {
-  list-style: none;
-  margin-bottom: 1rem;
-}
-
-.steps {
-  margin-top: 0.5rem;
-}
-
-/* Style for timer */
-div.timer {
-  align-items: center;
-  background-color: white;
-  border: 3px solid gold;
-  border-radius: 25px;
-  bottom: 0;
-  display: block;
-  float: left;
-  height: 140px;
-  left: 0;
-  margin: 1rem;
-  overflow: auto;
-  padding: 1rem;
-  position: fixed;
-  width: auto;
-}
-
-div.timer:hover {
-  cursor: grab;
-}
-
-/* stylelint-disable-next-line no-descending-specificity */
-div.timer button {
-  background-color: #fff0d4;
-  border-radius: 10px;
-}
-
-div.timeInput {
-  display: inline-flex;
-}
-
-p.showTimer {
-  font-size: 24px;
-  text-align: center;
-}
-
-button.setTime {
-  margin-left: 0.2rem;
-}
-
-::-webkit-input-placeholder { /* Chrome/Opera/Safari */
-  font-size: smaller;
-}
