--- demo-code/testHelper.html
@@ -1,47 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <title>Test html</title>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  </head>
-
-  <body>
-    <script type="module">
-      import * as helper from './apiHelpers.js';
-      import * as storage from './storage/storage.js'
-      
-      localStorage.clear();
-      //storeRecipeData Test: new category with two new recipe-------------
-      storage.storeRecipeData("cat1", [
-        { id:296687, title: "cat1_recipe1" },
-        { id:2, title: "cat1_recipe2" },
-      ]);
-      /*
-      helper.getRecipesByName('potato', 1)
-        .then((data) => {
-          console.log("Testing getRecipesByName");
-          console.log(data);
-      }); 
-      
-      helper.getRecipesByCuisine("mexican",1)
-        .then((data) => {
-          console.log("Testing getRecipesByCuisine");
-          console.log(data);
-      });
-      helper.getRecipesByType('lunch', 1)
-        .then((data) => {
-          console.log("Testing getRecipesByType");
-          console.log(data);
-      });
-      */
-      helper.getRecipesByAutocomplete('chick', 1)
-        .then((data) => {
-          console.log("Testing getRecipesByAutocomplete");
-          console.log(data);
-      });
-      
-    </script>
-  </body>
-</html>
